backend/requirements.txt
alembic==1.18.3
fastapi==0.128.3
uvicorn==0.40.0

SQLAlchemy==2.0.46
psycopg2-binary==2.9.11

pydantic==2.12.5
pydantic-settings==2.12.0
email-validator==2.3.0
python-dotenv==1.2.1
python-multipart==0.0.22

passlib==1.7.4
bcrypt==5.0.0
python-jose==3.5.0
cryptography==46.0.4

pandas==3.0.0
numpy==2.4.2
openpyxl==3.1.5
et_xmlfile==2.0.0

httpx==0.28.1

PyYAML==6.0.3

# Opcionales / runtime (no obligatorios, pero útiles)
watchfiles==1.1.1
websockets==16.0

# Opcionales de rendimiento (NO en Windows)
uvloop==0.22.1; platform_system != "Windows"
httptools==0.7.1; platform_system != "Windows"